requirements.txt
supabase==2.10.0
apify-client==1.8.1
google-cloud-language==2.13.4
openai==1.51.0
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn==0.31.0
